+++ apps/frontend/src/app/Data/Characters/Somnia/data_gen.json
