+++ apps/frontend/src/app/Data/Characters/Somnia/skillParam_gen.json
