--- libs/g-assets/src/gen/chars/Nekara/index.ts
@@ -1,34 +1,0 @@
-// This is a generated index file.
-import icon from './UI_AvatarIcon_Nekara.png'
-import iconSide from './UI_AvatarIcon_Side_Nekara.png'
-import banner from './UI_NameCardPic_Nekara_P.png'
-import bar from './UI_NameCardPic_Nekara_Alpha.png'
-import skill from './Skill_S_Nekara_01.png'
-import burst from './Skill_E_Nekara_01_HD.png'
-import passive1 from './UI_Talent_S_Nekara_05.png'
-import passive2 from './UI_Talent_S_Nekara_06.png'
-import passive3 from './UI_Talent_S_Nekara_07.png'
-import constellation1 from './UI_Talent_S_Nekara_01.png'
-import constellation2 from './UI_Talent_S_Nekara_02.png'
-import constellation3 from './Skill_E_Nekara_01_HD.png'
-import constellation4 from './UI_Talent_S_Nekara_03.png'
-import constellation5 from './Skill_S_Nekara_01.png'
-import constellation6 from './UI_Talent_S_Nekara_04.png'
-const data = {
-  icon,
-  iconSide,
-  banner,
-  bar,
-  skill,
-  burst,
-  passive1,
-  passive2,
-  passive3,
-  constellation1,
-  constellation2,
-  constellation3,
-  constellation4,
-  constellation5,
-  constellation6,
-} as const
-export default data
