+++ libs/g-assets/src/gen/chars/Somnia/Skill_E_Somnia_01_HD.png
