+++ libs/g-assets/src/gen/chars/Somnia/Skill_S_Somnia_01.png
