+++ libs/g-assets/src/gen/chars/Somnia/UI_AvatarIcon_Side_Somnia.png
