+++ libs/g-assets/src/gen/chars/Somnia/UI_AvatarIcon_Somnia.png
