+++ libs/g-assets/src/gen/chars/Somnia/UI_NameCardPic_Somnia_P.png
