+++ libs/g-assets/src/gen/chars/Somnia/index.ts
@@ -1,0 +1,34 @@
+// This is a generated index file.
+import icon from './UI_AvatarIcon_Somnia.png'
+import iconSide from './UI_AvatarIcon_Side_Somnia.png'
+import banner from './UI_NameCardPic_Somnia_P.png'
+import bar from './UI_NameCardPic_Somnia_Alpha.png'
+import skill from './Skill_S_Somnia_01.png'
+import burst from './Skill_E_Somnia_01_HD.png'
+import passive1 from './UI_Talent_S_Somnia_05.png'
+import passive2 from './UI_Talent_S_Somnia_06.png'
+import passive3 from './UI_Talent_S_Somnia_07.png'
+import constellation1 from './UI_Talent_S_Somnia_01.png'
+import constellation2 from './UI_Talent_S_Somnia_02.png'
+import constellation3 from './Skill_E_Somnia_01_HD.png'
+import constellation4 from './UI_Talent_S_Somnia_03.png'
+import constellation5 from './Skill_S_Somnia_01.png'
+import constellation6 from './UI_Talent_S_Somnia_04.png'
+const data = {
+  icon,
+  iconSide,
+  banner,
+  bar,
+  skill,
+  burst,
+  passive1,
+  passive2,
+  passive3,
+  constellation1,
+  constellation2,
+  constellation3,
+  constellation4,
+  constellation5,
+  constellation6,
+} as const
+export default data
